backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-multipart
python-dotenv
groq
